Rename to event-view.ts
