+++ src/lib/components/event/event-summary-row.svelte
@@ -2,7 +2,7 @@
   import Icon from 'svelte-fa';
   import { faAngleDown, faAngleUp } from '@fortawesome/free-solid-svg-icons';
 
-  import { eventFilterSort, eventShowElapsed } from '$lib/stores/event-views';
+  import { eventFilterSort, eventShowElapsed } from '$lib/stores/event-view';
   import { timeFormat } from '$lib/stores/time-format';
   import { getSingleAttributeForEvent } from '$lib/utilities/get-single-attribute-for-event';
