+++ src/lib/stores/event-view.ts
