+++ src/routes/namespaces/[namespace]/workflows/[workflow]/[run]/history/__layout.svelte
@@ -41,7 +41,7 @@
   import { routeForEventHistory } from '$lib/utilities/route-for';
   import { getWorkflowStartedAndCompletedEvents } from '$lib/utilities/get-started-and-completed-events';
   import { formatDate } from '$lib/utilities/format-date';
-  import { eventViewType } from '$lib/stores/event-views';
+  import { eventViewType } from '$lib/stores/event-view';
 
   import ToggleButton from '$lib/components/toggle-button.svelte';
   import ToggleButtons from '$lib/components/toggle-buttons.svelte';
